Add genhtml step
